Be accurate in disallowing sampling rates in SoftFM.
The rtl-sdr driver code disallows most of the sampling rates below
900 kS/s. However, it allows a small window of sampling rates
between [250001, 300000]. The update to the conditional is along
the lines of the conditions imposed by the RTL SDR driver here:

http://cgit.osmocom.org/rtl-sdr/tree/src/librtlsdr.c

The default sampling rate of 1 MS/s causes severe buffer underruns on
a Raspberry Pi 2. However choosing a sampling rate of 250 kS/s and
lower values for PCM sampling causes successful FM demodulation of
strong FM stations on the device.